Reorganise our functions a bit.
